post

+++ app/Http/Controllers/TasksController.php
@@ -11,13 +11,21 @@ class TasksController extends Controller
     // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        // タスク一覧を取得
-        $tasks = Task::all();
-
-        // タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        // 認証済みの場合
+        if (\Auth::check()) { 
+            // 認証済みユーザ（閲覧者）のみのタスク一覧を取得
+            $user = \Auth::user();
+            $tasks = $user->tasks;
+    
+            // タスク一覧ビューでそれを表示
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+        }
+        // 認証済みでない場合
+        else {
+            return view('tasks.index');
+        }
     }
 
     // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
@@ -41,10 +49,10 @@ class TasksController extends Controller
         ]);
         
         // タスクを作成
-        $task = new Task;
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -56,10 +64,15 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
 
-        // タスク詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        // 認証済みユーザ（閲覧者）がそのタスクの所有者である場合は、タスク詳細ビューでそれを表示
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }
+        else {
+            return redirect('/');
+        }
     }
 
     // getでtasks/id/editにアクセスされた場合の「更新画面表示処理」
@@ -68,10 +81,15 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
 
-        // タスク編集ビューでそれを表示
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        // 認証済みユーザ（閲覧者）がそのタスクの所有者である場合は、タスク編集ビューでそれを表示
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }
+        else {
+            return redirect('/');
+        }
     }
 
     // putまたはpatchでtasks/idにアクセスされた場合の「更新処理」
@@ -85,10 +103,12 @@ class TasksController extends Controller
         
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // タスクを更新
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+        // 認証済みユーザ（閲覧者）がそのタスクの所有者である場合は、タスクを更新
+        if (\Auth::id() === $task->user_id) {
+            $task->status = $request->status;    // 追加
+            $task->content = $request->content;
+            $task->save();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -99,9 +119,12 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
-
+        
+        // 認証済みユーザ（閲覧者）がそのタスクの所有者である場合は、タスクを削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
         // トップページへリダイレクトさせる
         return redirect('/');
     }
